Move View Test Date to components directory
